CRUD authentication

+++ Backend/Controllers/goalControllers.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../Models/goalModels.js')
+const User = require('../Models/userModels.js')
 
 //Description: Get goals
 //Route: GET /api/goals
 //Access: Private
 const getGoals = asyncHandler( async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find( { user: req.user._id })
 
     res.status(200).json(goals);
 })
@@ -22,8 +23,10 @@ const setGoals = asyncHandler( async (req, res) => {
         throw new Error('Please enter a goal')
     }
 
+    //!ADD HERE MAYBE ROLE
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user._id
     })
 
     res.status(200).json(goal);
@@ -40,6 +43,20 @@ const updateGoal = asyncHandler( async (req, res) => {
     if (!goal) {
         res.status(404)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if (!user) {
+        res.status(404)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the user that created the goal
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('Not authorized to update goal')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(
@@ -66,6 +83,21 @@ const deleteGoal = asyncHandler( async (req, res) => {
       res.status(400)
       throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if (!user) {
+        res.status(404)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the user that created the goal
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('Not authorized to delete goal')
+    }
+
 
     await goal.deleteOne()
